Move expression-input and toggle-slider to editor-common module. Move basic-input-section to tool-editor module.
