+++ src/app/editor-common/components/toggle-slider/toggle-slider.component.scss
