+++ src/app/tool-editor/object-inspector/basic-section/basic-input-section.component.scss
